LayoutTests/storage/indexeddb/transaction-event-propagation.html

<html>
<head>
<link rel="stylesheet" href="../../fast/js/resources/js-test-style.css">
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="../../fast/js/resources/js-test-post-function.js"></script>
<script src="resources/shared.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Test event propogation on IDBTransaction.");
if (window.layoutTestController) 
    layoutTestController.waitUntilDone();

function test()
{
    result = evalAndLog("webkitIndexedDB.open('name')");
    verifyResult(result);
    result.onsuccess = setVersion;
    result.onerror = unexpectedErrorCallback;
}

function setVersion()
{
    db = evalAndLog("db = event.result");

    result = evalAndLog("db.setVersion('new version')");
    verifyResult(result);
    result.onsuccess = deleteExisting;
    result.onerror = unexpectedErrorCallback;
}

function deleteExisting()
{
    window.trans = evalAndLog("trans = event.result");
    shouldBeTrue("trans !== null");
    trans.onabort = unexpectedAbortCallback;
    evalAndLog("trans.oncomplete = startTest");

    deleteAllObjectStores(db, createObjectStore);
}

function createObjectStore()
{
    store = evalAndLog("store = db.createObjectStore('storeName', null)");
    result = evalAndLog("store.add({x: 'value', y: 'zzz'}, 'key')");
    result.onerror = unexpectedErrorCallback;
}

function startTest()
{
    debug("Verifing abort");
    trans = evalAndLog("trans = db.transaction([], webkitIDBTransaction.READ_WRITE)");
    evalAndLog("trans.onabort = abortFiredCallback");
    evalAndLog("trans.oncomplete = unexpectedAbortCallback");
    evalAndLog("db.addEventListener('abort', dbAbortCaptureCallback, true)");
    evalAndLog("db.addEventListener('abort', dbAbortBubbleCallback, false)");
    evalAndLog("db.addEventListener('complete', unexpectedCompleteCallback, true)");
    evalAndLog("db.addEventListener('complete', unexpectedCompleteCallback, false)");
    store = evalAndLog("store = trans.objectStore('storeName')");
    evalAndLog("store.add({x: 'value', y: 'zzz'}, 'key')");
    dbCaptureFired = false;
    abortFired = false;
    dbBubbleFired1 = false;
}

function dbAbortCaptureCallback()
{
    debug("");
    debug("In IDBDatabase abort capture");
    shouldBeFalse("dbCaptureFired");
    shouldBeFalse("abortFired");
    shouldBeFalse("dbBubbleFired1");
    dbCaptureFired = true;
}

function abortFiredCallback()
{
    debug("");
    debug("In abort handler");
    shouldBeTrue("dbCaptureFired");
    shouldBeFalse("abortFired");
    shouldBeFalse("dbBubbleFired1");
    abortFired = true;
}

function dbAbortBubbleCallback()
{
    debug("");
    debug("In IDBDatabase error bubble");
    shouldBeTrue("dbCaptureFired");
    shouldBeTrue("abortFired");
    shouldBeFalse("dbBubbleFired1");
    dbBubbleFired1 = true;
    debug("");
    debug("Verifing success.");
    trans = evalAndLog("trans = db.transaction([], webkitIDBTransaction.READ_WRITE)");
    evalAndLog("trans.oncomplete = completeFiredCallback");
    evalAndLog("trans.onabort = unexpectedAbortCallback");
    evalAndLog("db.removeEventListener('abort', dbAbortCaptureCallback, true)");
    evalAndLog("db.removeEventListener('abort', dbAbortBubbleCallback, false)");
    evalAndLog("db.removeEventListener('complete', unexpectedCompleteCallback, true)");
    evalAndLog("db.removeEventListener('complete', unexpectedCompleteCallback, false)");
    evalAndLog("db.addEventListener('complete', dbCompleteCaptureCallback, true)");
    evalAndLog("db.addEventListener('complete', dbCompleteBubbleCallback, false)");
    evalAndLog("db.addEventListener('abort', unexpectedAbortCallback, true)");
    evalAndLog("db.addEventListener('abort', unexpectedAbortCallback, false)");
    store = evalAndLog("store = trans.objectStore('storeName')");
    evalAndLog("store.add({x: 'value', y: 'zzz'}, 'key2')");
    dbCaptureFired = false;
    completeFired = false;
    dbBubbleFired2 = false;
}

function dbCompleteCaptureCallback()
{
    debug("");
    debug("In IDBDatabase complete capture");
    shouldBeFalse("dbCaptureFired");
    shouldBeFalse("completeFired");
    shouldBeFalse("dbBubbleFired2");
    dbCaptureFired = true;
}

function completeFiredCallback()
{
    debug("");
    debug("In IDBRequest handler");
    shouldBeTrue("dbCaptureFired");
    shouldBeFalse("completeFired");
    shouldBeFalse("dbBubbleFired2");
    completeFired = true;
    debug("");
    done();
}

function dbCompleteBubbleCallback()
{
    debug("");
    debug("In IDBDatabase complete bubble");
    testFailed("Complete bubble callback should NOT fire");
    dbBubbleFired2 = true;
}

test();

var successfullyParsed = true;

</script>
</body>
</html>
